% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayMeans.R
\name{arrayMean}
\alias{arrayMean}
\title{Truncated normal distribution}
\usage{
arrayMean(y)
}
\arguments{
\item{y}{A numeric array.}
}
\value{
A numeric vector representing the mean of means with length equal to
\code{ncol(y)}.
}
\description{
This function calculates means across an array's third dimension, then its
second dimension.
}
\examples{
a <- matrix(rnorm(5 * 4, mean = 0, sd = 1), 5, 4)
b <- matrix(rnorm(5 * 4, mean = 1, sd = 1), 5, 4)
c <- matrix(rnorm(5 * 4, mean = 2, sd = 1), 5, 4)
y <- sapply(list(a, b, c), identity, simplify="array")
arrayMean(y)
}
