% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataframe.R
\name{buildDataAgg}
\alias{buildDataAgg}
\title{Build output dataframes for aggregate data}
\usage{
buildDataAgg(dirNames, agVars, keyVarName)
}
\arguments{
\item{dirNames}{A character vector representing directories, each of which
contains multiple scenarios nested within subdirectories.}

\item{agVars}{A character vector representing aggregate performance metrics
to include in the dataframe.}

\item{keyVarName}{A character vector specifying which variable differs among
scenarios within a subdirectory within a directory (e.g. fixed exploitation
rate).}
}
\value{
Returns a dataframe with columns for the keyVariable, management
procedure, operating model, plotOrder, variable, its median value, its lower
quantile, and its upper quantile.
}
\description{
This is one of two sister functions that are extensions of
\code{genOutputList} and used to builds dataframes around key variables that
change across scenarios. One is for CU-specific data and the other for
aggregate data. Each combination of directory, subdirectory, and key
variable levels should represent a unique scenario (i.e. they can vary across
up to three dimensions). For example, different key variables (proportion
allocated to mixed fisheries) nested within OM (productivity regime) nested
within higher level MP or additional OM (e.g. fixed exploitation or TAM
rule).
}
\examples{
TO BE COMPLETED

}
