% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataframe.R
\name{buildDataCU}
\alias{buildDataCU}
\title{Build output dataframes for CU data}
\usage{
buildDataCU(dirNames, cuVars, keyVarName, selectedCUs = NULL)
}
\arguments{
\item{dirNames}{A character vector representing directories, each of which
contains multiple scenarios nested within subdirectories.}

\item{cuVars}{A character vector representing CU-specific performance metrics
to include in the dataframe.}

\item{keyVarName}{A character vector specifying which variable differs among
scenarios within a subdirectory within a directory (e.g. fixed exploitation
rate).}

\item{selectedCUs}{An optional character vector of CU names used to subset
output data if CU-specific data are being pulled (i.e. \code{agg = FALSE}).}
}
\value{
Returns a dataframe with columns for the keyVariable, management
procedure, operating model, plotOrder, cuName, muName, variable, its median
value, its lower quantile, and its upper quantile.
}
\description{
This is one of two sister functions that are extensions of
\code{genOutputList} and used to builds dataframes around key variables that
change across scenarios. One is for CU-specific data and the other for
aggregate data. Each combination of directory, subdirectory, and key
variable levels should represent a unique scenario (i.e. they can vary across
up to three dimensions). For example, different key variables (proportion
allocated to mixed fisheries) nested within OM (productivity regime) nested
within higher level MP or additional OM (e.g. fixed exploitation or TAM
rule).
}
