% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCanEROU_fixedER.R
\name{calcCanEROU_fixedER}
\alias{calcCanEROU_fixedER}
\title{Title}
\usage{
calcCanEROU_fixedER(canER, cvERSMU, randomVar = T, maxER)
}
\arguments{
\item{canER}{A numeric representing the target Canadian exploitation rate}

\item{cvERSMU}{A numeric representing annual variability in ER}

\item{randomVar}{A TRUE/FALSE variable indicating whether the Canadian ER should have
annual implementation error around the target}
}
\value{
Returns a four element list of numeric vectors with length equal to
forecast:Total Canadian TAC, Canadian mixed fishery TAC, Canadian single fishery TAC, American TAC
}
\description{
Title
}
\examples{
calcCanEROU_fixedER(canER=0.3, cvERSMU=0.1)


}
