% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcObservedError.R
\name{calcErr}
\alias{calcErr}
\title{Calculate observed error}
\usage{
calcErr(obs, true)
}
\arguments{
\item{obs}{A numeric vector representing observed data.}

\item{true}{A numeric vector representing true data.}
}
\value{
Returns a numeric vector of absolute relative error.
}
\description{
This helper function calculates relative error rates with a few corrections
to account for missing or wonky data.
}
