% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcObsCatch.R
\name{calcObsCatch}
\alias{calcObsCatch}
\title{Calculate observed catch}
\usage{
calcObsCatch(
  catchVec,
  recVec,
  manUnit,
  tauCatch,
  stkID,
  catchObsErr,
  extinctThresh
)
}
\arguments{
\item{catchVec}{A numeric vector representing realized CU-specific catches
(i.e. outcome uncertainty incorporated).}

\item{recVec}{A numeric vector representing CU-specific recruits (RY).}

\item{manUnit}{A character vector identifying the MU that each CU belongs to.}

\item{tauCatch}{A numeric representing multivariate logistic error associated
with assigning catch in mixed stock fisheries to the correct CU.}

\item{stkID}{A character vector of CU names.}

\item{catchObsErr}{A numeric representing log-nomrally distributed error in
catch observations.}

\item{extinctThresh}{A numeric representing the extinction threshold for the
aggregate.}
}
\value{
Returns a numeric vector of observed catch.
}
\description{
This function adjusts realized catches by normally distributed observation
error to simulate observed catches. For mixed stock fisheries a tau
parameter representing multivariate logistic error is also incorporated to
account for uncertainty in the stock assignment process.
}
\examples{
#Note that the function is intended to receive vectors rather than the DFs
#used in this example to increase efficiency.
head(exampleHCRList)

catch <- exampleHCRList$mixCatch
rec <- exampleHCRList$recRY
manUnit <- exampleHCRList$mu
stock <- exampleHCRList$stock
calcObsCatch(catch, rec, manUnit, tauCatch = 0.1, stock, catchObsErr = 0.2,
extinctThresh = 0.0001)
}
