% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRealCatch.R
\name{calcRealCatch}
\alias{calcRealCatch}
\title{Calculate realized catches}
\usage{
calcRealCatch(rec, tac, sigma = 0.1, random = FALSE, setSeedInput = NULL)
}
\arguments{
\item{rec}{A numeric vector of length nCU representing CU-specific
recruitment.}

\item{tac}{A numeric vector of length nCU representing CU-specific target
total allowable catch rates (generally passed from \code{calcTAC} function).}

\item{sigma}{A numeric representing the standard deviation of the shape
parameter.}

\item{random}{A logical (default `FALSE`) used to restandardize random number
generator because `rbeta()` seems to produce variable draws (should be
corrected in future).}

\item{setSeedInput}{A numeric representing the combination MC trial and
simulation year to ensure that scenarios are sampling the seed, but allowing
draws to otherwise vary.}
}
\value{
Returns a vector of length nCU representing realized catches
}
\description{
This function calculates realized catch based on beta distributed outcome
uncertainty.
}
\details{
Outcome uncertainty is often easier to parameterize using observed deviations
in target and realized exploitation rates rather than catches. Here target
total allowable catches are back-converted to exploitation rates and realized
catches are generated using  beta distributed outcome uncertainty (as in
Anderson et al. 2015 J. App. Ecol.).

For simplicity's sake when TAC is zero, realized HR is zero. Therefore to
represent "significant" catch even when TAC is zero, pass large sigma values
and a small, non-zero TAC value.

**Note** the sigma values representing the standard deviation of the shape
parameter for the beta distribution are **not** equivalent to the SD of a
normal distribution. Values larger than ~0.2 can result in U-shaped realized
exploitation rates. Distribution should be parameterized based on data or use
default value (0.1) from Pestes et al. 2008 for Cultus Lake sockeye salmon.

**Note** small target harvest rates and/or high sigmas can produce negative
location parameters which result in NAs; replace with small values
(see betaDistributionBounds.Rmd for details) .
}
\examples{
head(exampleHCRList)
exRec <- exampleHCRList$recRY
exTAC <- 0.2 * exRec
calcRealCatch(exRec, exTAC, sigma = 0.1)

}
