% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTAC_fixedER.R
\name{calcTAC_fixedER}
\alias{calcTAC_fixedER}
\title{Title}
\usage{
calcTAC_fixedER(
  rec,
  canER,
  amER,
  ppnMixVec,
  cvERcan,
  cvERam,
  randomVar = T,
  runif = NULL,
  maxER
)
}
\arguments{
\item{rec}{A numeric representing MU-specific return abundance}

\item{canER}{A numeric representing the target Canadian exploitation rate}

\item{amER}{A numeric representing the target American exploitation rate}

\item{ppnMixVec}{A numeric representing the proportion of the Canadian TAC
allocated to mixed stock fisheries}

\item{cvERcan}{A numeric representing annual variability in domestic ER}

\item{cvERam}{A numeric representing annual variability in US interceptions}

\item{randomVar}{A TRUE/FALSE variable indicating whether the Canadian ER should have
annual implementation error around the target}

\item{runif}{A vector of random numbers of length equal to the number of CUs
which provides constant CU-specific deviations in ER from annual values}
}
\value{
Returns a four element list of numeric vectors with length equal to
forecast:Total Canadian TAC, Canadian mixed fishery TAC, Canadian single fishery TAC, American TAC
}
\description{
Title
}
\examples{
calcTAC_fixedER(rec=1000, canER=0.3, amER = 0.1, ppnMix = 0.5)


}
