% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCycle.R
\name{checkInteger}
\alias{checkInteger}
\title{Check integer}
\usage{
checkInteger(x)
}
\arguments{
\item{x}{A numeric.}
}
\value{
Returns a logical confirming whether x is or is not an integer.
}
\description{
Helper function to calculate which cycle line an observed year belongs to. Used in
\code{genCygle}.
}
\examples{
checkInteger(3.4)
checkInteger(3)

}
