% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{cv}
\usage{
cv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns a numeric
}
\description{
Simple helper function calculate coefficient of variation.
}
\examples{
x <- rnorm(50, mean = 0, sd = 1)
cv(x)
}
