% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinctionCheck.R
\name{extinctionCheck}
\alias{extinctionCheck}
\title{Extinction check}
\usage{
extinctionCheck(y, gen, extinctThresh, spwnMat)
}
\arguments{
\item{y}{A numeric representing the current year within the simulation
(i.e. has a quasi-extinction occurred based on abundance in y-gen:y).}

\item{gen}{A numeric representing generation length for the species being
modeled.}

\item{extinctThresh}{A numeric representing the assumed quasi-extinction
threshold.}

\item{spwnMat}{A matrix of spawner abundances with ncol = nCU and minimum
nrow = gen.}
}
\value{
Returns a binary vector of length = ncol(spwnMat) with ones
representing extinct stocks/CUs.
}
\description{
This function checks whether a population has gone quasi-extinct based on
spawner abundances, generation length, and the extinction threshold.
}
\examples{
nCUs <- 5
gen <- 4
yr <- 6
sMat <- matrix(0, nrow = yr, ncol = nCUs)
for (i in 1:nCUs) {
  sMat[ , i] <- exp(rnorm(6, mean = 0.5, sd = 1))
}
sMat[ , 5] <- c(0.1, 0.0001, 0.0001, 0.0001, 0.0001, 0.0001)
extinctionCheck(y = yr, gen = gen, extinctThresh = 0.001, spwnMat = sMat)
}
