% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLm}
\alias{fastLm}
\title{Estimate coefficients for linear model.}
\usage{
fastLm(y, X)
}
\arguments{
\item{y}{A numeric vector of response variables}

\item{X}{A numeric matrix of predictor variables}
}
\value{
A numeric list with estimated coefficients and standard deviations.
}
\description{
This is a function to rapidly fit linear models and return estimated
coefficients.
}
