% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCycle.R
\name{genCycle}
\alias{genCycle}
\title{Generate cycle lines}
\usage{
genCycle(firstObs, simLength)
}
\arguments{
\item{firstObs}{A numeric representing the first year of observed data.}

\item{simLength}{A numeric representing the length of the simulation period.}
}
\value{
Returns a numeric vector representing cycle line IDs for output data
with \code{length(simLength)}.
}
\description{
This function generates cycle lines necessary to model Fraser River stocks
exhibiting delayed density dependence (e.g. Lower Shuswap). Note that it is
not a universal function - it can only generate four year cycles and assumes
that the observed data begins after 1900.
}
\details{
Depends on the helper function \code{checkInteger}.
}
\examples{
genCycle(firstObs = 2009, simLength = 50)

}
