% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationMean.R
\name{genMean}
\alias{genMean}
\title{Generational mean}
\usage{
genMean(vec, gen)
}
\arguments{
\item{vec}{A numeric vector, typically of spawner abundances.}

\item{gen}{A numeric representing generation length (years).}
}
\value{
Returns a numeric vector of the running mean
}
\description{
Calculate running mean of a vector over generation length.
}
\examples{
rec <- recMatrix[ , 1]
genMean(rec, gen = 4)

}
