% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutputList.R, R/makeOutputList.R
\name{genOutputList}
\alias{genOutputList}
\title{Generate list of outputs}
\usage{
genOutputList(dirName, subDirName = NULL, selectedCUs = NULL,
  agg = TRUE, aggTS = FALSE)

genOutputList(dirName, subDirName = NULL, selectedCUs = NULL,
  agg = TRUE, aggTS = FALSE)
}
\arguments{
\item{dirName}{A character vector representing the directory to pull data
from.}

\item{subDirName}{An optional character vector used to specify a single
subdirectory within \code{dirName} to pull data from.}

\item{selectedCUs}{An optional character vector of CU names used to subset
output data if CU-specific data are being pulled (i.e. \code{agg = FALSE}).}

\item{agg}{A logical stating whether aggregate (\code{TRUE}; default) or
CU-specific (\code{FALSE}) data should be pulled.}

\item{aggTS}{A logical stating whether aggregate summary performance metric
data (\code{FALSE}; default) or aggregate time series data (\code{TRUE})
should be pulled.}

\item{dirName}{A character vector describing directory where output data saved.}

\item{subDirName}{A character vector describing subdirectory where output data saved.
Defaults to NULL.}

\item{selectedCUs}{A character vector of CUs that will be subsetted from aggregate. Defaults
to NULL.}

\item{agg}{Is the generated list summarizing aggregate or CU-specific data? Defaults to TRUE.}

\item{aggTS}{If agg == TRUE, is the generated list summarizing aggregate time series data?
Defaults to FALSE.}
}
\value{
Returns a list of output data of length equal to the number of unique
scenarios within a directory.

Returns a list of subdirectory outputs within the directory.
}
\description{
This function pulls data generated by \code{recoverySim} from a given
directory and binds it into a list. Each first order list element consists
of a single scenario (i.e. one OM and one MP). The second order list
elements depend on whether CU-specific, aggregate, or aggregate time series
data are extracted.

This function is used to generate lists of summary data based on the outputs of recoverySim.
}
\details{
This is a function specific to the samSim package. Generally a directory represents either a
single operating model/management procedure with subdirectories representing representing OM/MPs
nested within creating multiple unique scenarios.
}
\examples{
TO BE COMPLETED


}
