% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infillMatrices.R
\name{geoMean}
\alias{geoMean}
\title{Calculate geometric mean}
\usage{
geoMean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns a numeric equal to geometric mean
}
\description{
Helper function to calculate geometric means. Used in \code{infill}.
}
\examples{
x <- c(2, 5, 12, 9)
geoMean(x)
}
