% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Gather stock-recruit residuals}
\usage{
getResiduals(recList, modelVec)
}
\arguments{
\item{recList}{A list of length nCU with each element a dataframe with nrow
= to length of the maximum observed timeseries. Each dataframe should
contain the following columns: an estimate of total recruitment (by brood
year), spawner abundance, and brood year. Lagged spawner abundances are
automatically generated by the function.}

\item{modelVec}{A character vector of model names with length nCU
(\code{ricker} or \code{larkin}).}
}
\value{
A numeric matrix of model residuals with nrow = nrow(input data)
and ncol = nCU.
}
\description{
This function fits simple stock-recruit models (Ricker or Larkin) to
generate model residuals that can be merged with simulated recruitment
deviations. Useful for posterior predictive comparisons and generating
output figures if model residuals are a variable of interest.
}
\examples{
EXAMPLE DATA NEEDS TO BE GENERATED
}
