% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRPars.R
\name{getSRPars}
\alias{getSRPars}
\title{Summarize stock-recruit parameters}
\usage{
getSRPars(pars, alphaOnly = TRUE, highP = 0.9, lowP = 0.1, stks = NULL)
}
\arguments{
\item{pars}{A dataframe containing stock-recruit parameters. Should be
formatted so that stock/population (\code{stk}) is in first column and each
subsequent column contains one parameter; each row represents one posterior
sample.}

\item{alphaOnly}{A logical argument. If TRUE then percentiles are only used
for alpha parameter, all others use medians. If FALSE then percentiles are
calculated and returned for all parameters.}

\item{highP, lowP}{Numerics that represent high and low percentiles that are
being estimated.}

\item{stks}{A vector that is used to identify a subset of stocks for which
parameters are returned. Can be either a numeric or character, but should
match the format of the first column of the input dataframe.}
}
\value{
Returns a list of length three (\code{pLow, pMed, pHigh}) containing
lower percentiles, medians and higher percentiles, respectively. Note that
if alphaOnly = TRUE, the same median values will be returned for non-alpha
parameters in all list elements.
}
\description{
This function samples relevant stock-recruit parameters from a dataframe
containing posterior samples from MCMC sampling. Sampling is based on
percentile arguments.
}
\examples{
head(rickerParameters)
d <- getSRPars(rickerParameters, alphaOnly = FALSE)
head(d$pLow) #lower percentile estimates
head(d$pMed) #median estimates

d <- getSRPars(larkinParameters, alphaOnly = TRUE)
head(d$pLow) #lower percentile estimates for alpha only
}
