% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRPars_randomSamp.R
\name{getSRPars_randomSamp}
\alias{getSRPars_randomSamp}
\title{Sample stock-recruit parameters from posterior distribution}
\usage{
getSRPars_randomSamp(pars, stks = NULL)
}
\arguments{
\item{pars}{A dataframe containing stock-recruit parameters. Should be
formatted so that stock/population (\code{stk}) is in first column and each
subsequent column contains one parameter; each row represents one posterior
sample.}

\item{stks}{A vector that is used to identify a subset of stocks for which
parameters are returned. Can be either a numeric or character, but should
match the format of the first column of the input dataframe.}
}
\value{
Returns a data frame with SR parameters for each stock that come from a single, randomly
sampled MCMC replicate.Each row in the data frame represents a stock.
}
\description{
This function samples relevant stock-recruit parameters from a dataframe
containing posterior samples from MCMC sampling. Sampling among MCMC replicates is
random with replacement.
}
\examples{
head(rickerParameters)
d <- getSRPars(rickerParameters, alphaOnly = FALSE)
head(d$pLow) #lower percentile estimates
head(d$pMed) #median estimates

d <- getSRPars(larkinParameters, alphaOnly = TRUE)
head(d$pLow) #lower percentile estimates for alpha only
}
