% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverySimulator.R
\name{here}
\alias{here}
\title{Recovery simulator}
\usage{
here(...)
}
\arguments{
\item{y}{TO BE DEFINED}
}
\value{
TO BE DEFINED.
}
\description{
Closed-loop simulation tool to assess management procedures and inform
		 Pacific salmon rebuilding strategies;
		 Primarily based on C. Holt's CSAS chum model; earlier model version (SimpleSimWCatch.R Feb 7) saved in scripts/Defunct
			 Simulation runs primed with observed SR data (recDat input)
		 Includes data generation, variation in age structure, survey design, and variable exploitation rules
	 (constant ER);
	 Uses Ricker formulation \code{R=S(exp(a-bS))};
	 Additional details in changesToChumModel.md in reports directory
}
