% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockRecruitModels.R
\name{larkinModel}
\alias{larkinModel}
\title{Generate recruit abundance with Larkin model}
\usage{
larkinModel(S, Sm1, Sm2, Sm3, a, b, b1, b2, b3, error)
}
\arguments{
\item{S}{A numeric vector of spawner abundances.}

\item{Sm1, Sm2, Sm3}{A numeric vector of spawner abundances at 1, 2 and 3 year
lags, respectively.}

\item{a}{A numeric vector of alpha values, i.e. productivity at low spawner
abundance.}

\item{b}{A numeric vector of beta values, i.e. density dependence para-
meter.}

\item{b1, b2, b3}{A numeric vector of delayed density dependent effects at 1,
2, and 3 yera lags, respectively.}

\item{error}{A numeric vector of recruitment deviations, typically generated
using \code{rmvnorm()} and relevant process variance estimates (sigma).}
}
\value{
Returns a numeric representing recruit abundance.
}
\description{
This function calculates recruitment from Larkin model (according to
Peterman et al. 2003, modified to take more recent parameter-
ization). Uses parameters in log space, like rickerMod, with
multivariate normally distributed errors, but cannot incorporate AR1 process
error because such models have not been validated.
}
\section{Note}{
 the log-normal bias correction has not been fixed for the
Larkin model.
}

\examples{
#Spawner and recruit values represent millions of fish, parameters
approximate those of Shuswap CU
larkinModel(S = 1.1, Sm1 = 0.4, Sm2 = 0.2, Sm3 = 0.15, a = 2.2, b = 0.29,
b1 = 0.42, b2 = 0.31, b3 = 0.21, error = 0.3)

}
