% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateWhen.R
\name{mutate_when}
\alias{mutate_when}
\title{mutate_when}
\usage{
mutate_when(data, ...)
}
\arguments{
\item{data}{Dataframe that meets standard dplyr requirements.}
}
\value{
Returns a dataframe subsetted as requested.
}
\description{
Simple helper function to allow conditional application of mutate.
}
\examples{
mtcars \%>\% mutate_when(
   mpg > 22,    list(cyl = 100),
   disp == 160, list(cyl = 200)
)
}
