% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{negll_schnute}
\alias{negll_schnute}
\title{negative log likelihood function for multivariate logistic likelihood}
\usage{
negll_schnute(pars, props)
}
\arguments{
\item{pars}{parameters to be estimates (initial guesses) mean proportion at age and variance. 
Make sure vector is compatible with observed proportions}

\item{props}{matrix of observed proportions at age}
}
\value{
negative log likelihood value
}
\description{
This function provides MLE estimates for the multivariate logistic likelihood
function describe by Schunute and Richards 1995. Originally coded by Paul van Dam-Bates
}
