% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDots.R
\name{plotAgDot}
\alias{plotAgDot}
\title{Aggregate dot plot}
\usage{
plotAgDot(
  agDat,
  group = "om",
  legendLab = "Operating\\nModel",
  xLab = NULL,
  yLab = NULL,
  plotTitle = NULL,
  axisSize = 14,
  dotSize = 4,
  lineSize = 1,
  legendSize = 14
)
}
\arguments{
\item{agDat}{Dataframe generated by \code{buildCUDat}.}

\item{group}{A character value that can take the values: \code{"mp", "om"}
and specifies along which categorical variable dot plots should be grouped.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{plotTitle}{A character representing the main plot title.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates dot plots with credible intervals for performance
metrics. Note that for aggregate data, variables should be filtered from
dataframe before it is passed as an argument (i.e. plots all variables by
default).
}
\examples{
trimDat <- agPlottingDF \%>\%
dplyr::filter(var \%in\% c("medSpawners", "medCatch", "ppnCULower", "ppnCUExtinct"))

plotAgDot(trimDat, group = "om",
         legendLab = "Secondary MPs", xLab = "Proportion Mixed Stock TAC",
         yLab = "Median Catch", axisSize = 14, dotSize = 4, lineSize = 1,
         legendSize = 14)

}
