% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTradeoffs.R
\name{plotAgTradeoff}
\alias{plotAgTradeoff}
\title{Aggregate trade off plot}
\usage{
plotAgTradeoff(agDat, consVar = "medSpawners", catchVar = "medCatch",
  facet = "om", showUncertainty = FALSE, legendLab = NULL,
  xLab = NULL, yLab = NULL, axisSize = 14, dotSize = 4,
  lineSize = 1.25, legendSize = 14, freeY = TRUE)
}
\arguments{
\item{agDat}{Dataframe generated by \code{buildAgDat}.}

\item{consVar}{A character value corresponding to a conservation based PM
in cuDat$vars.}

\item{catchVar}{A character value corresponding to a catch based PM in
cuDat$vars.}

\item{facet}{A character value that can take the values: \code{"mp", "om"}
and specifies along which categorical variable the plot should be faceted.}

\item{showUncertainty}{A logical specifying whether whiskers for each
variables credible interval should be plotted.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{freeY}{A logical specifying whether y-axis values should be allowed to
vary across facets.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates Kobe style plots for a series of conservation- and
catch-based performance metrics.
}
\examples{
TO BE ADDED ONCE EXAMPLE DATA PROVIDED

}
