% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDots.R
\name{plotCUDot}
\alias{plotCUDot}
\title{CU-specific dot plot}
\usage{
plotCUDot(
  cuDat,
  plotVar,
  group = "om",
  legendLab = NULL,
  xLab = NULL,
  yLab = NULL,
  plotTitle = NULL,
  axisSize = 14,
  dotSize = 4,
  lineSize = 1,
  legendSize = 14
)
}
\arguments{
\item{cuDat}{Dataframe generated by \code{buildCUDat}.}

\item{plotVar}{A character value corresponding to PM in cuDat$vars (can be
either catch- or conservation-based).}

\item{group}{A character value that can take the values: \code{"mp", "om"}
and specifies along which categorical variable dot plots should be grouped.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{plotTitle}{A character representing the main plot title.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates dot plots with credible intervals for performance
metrics.
}
\examples{

plotCUDot(cuPlottingDF, plotVar = "medCatch", group = "om",
         legendLab = "Secondary MPs", xLab = "Proportion Mixed Stock TAC",
         yLab = "Median Catch", axisSize = 14, dotSize = 4, lineSize = 1,
         legendSize = 14)

}
