% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTradeoffs.R
\name{plotCUTradeoff}
\alias{plotCUTradeoff}
\title{CU-specific trade off plot}
\usage{
plotCUTradeoff(cuDat, consVar = "medSpawners", catchVar = "medCatch",
  facet = "mu", panel = "om", showUncertainty = FALSE,
  legendLab = NULL, xLab = NULL, yLab = NULL, main = TRUE,
  axisSize = 14, dotSize = 4, lineSize = 1.25, legendSize = 14)
}
\arguments{
\item{cuDat}{Dataframe generated by \code{buildCUDat}.}

\item{consVar}{A character value corresponding to a conservation based PM
in cuDat$vars.}

\item{catchVar}{A character value corresponding to a catch based PM in
cuDat$vars.}

\item{facet}{A character value that can take the values:
\code{"mu", "cu", "mp", "om"} and specifies along which categorical variable
the plot should be faceted.}

\item{panel}{A character value that can take the values:
\code{"mu", mp", "om"} and specifies along which categorical variable new
pages in the output PDF will be generated.}

\item{showUncertainty}{A logical specifying whether whiskers for each
variable's credible interval should be plotted.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{main}{A logical specifying whether a plot title should be added.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates Kobe style plots for a series of conservation- and
catch-based performance metrics.
}
\examples{
TO BE ADDED ONCE EXAMPLE DATA PROVIDED

}
