% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagCU}
\alias{plotDiagCU}
\title{Diagnostic plot}
\usage{
plotDiagCU(
  varArray,
  varNames,
  stkName = NULL,
  model,
  ricB,
  larBList = NULL,
  medAbundance = NULL,
  nPrime,
  extinct,
  focalCU = NULL
)
}
\arguments{
\item{varArray}{A numeric array containing CU-specific data.}

\item{varNames}{A character vector containing the names of each variable
passed to the function (i.e. names for the third dimension of
\code{varArray}).}

\item{stkName}{An optional character vector containing stock/CU names.
Recommended for plotting.}

\item{model}{A character vector specifying whether each CU is fit to a
Ricker or Larkin model. Length equal to the number of CUs.}

\item{ricB}{A numeric vector of each CU's Ricker beta parameter.}

\item{larBList}{A numeric list of four numeric vectors, each representing
CU-specific estimates of lagged density dependent effects from Larkin models.}

\item{medAbundance}{A numeric matrix of each CU's median abundance plus upper
and lower quantiles.}

\item{nPrime}{Numeric representing length of pre-simulation priming period
(i.e. observation period) in the Monte Carlo trial.}

\item{extinct}{A binary vector of extinct (\code{1}) and extant (\code{0})
CUs.}

\item{focalCU}{A numeric identifying the CU that should be selected for
presentation. If \code{NULL} (default), a CU is selected at random.}
}
\value{
Returns a multi-page panel, multi-page plot.
}
\description{
This function generates diagnostic figures showing CU-specific traits from
one Monte Carlo trial within the recoverySimulator function. First figure is
a time series of spawner and recruit abundance through the observation and
simulation period for one CU (selected at random or with \code{focalCU}
argument). Second figure shows the stock-recruit relationship for that same
CU separating abundances from process (filled circles) and observation model
(closed circles). Remaining figures are time series from up to six CUs (if
more are passed, 6 will be selected at random) showing the metrics passed in
the input array.
}
