% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRadar.R
\name{plotRadar}
\alias{plotRadar}
\title{Radar plot}
\usage{
plotRadar(
  dat,
  xLab,
  plotVars = NULL,
  groupingVar = NULL,
  cu = FALSE,
  mainLab = NULL,
  legendLab = NULL,
  axisSize = 13
)
}
\arguments{
\item{xLab}{A character representing the x axis label.}

\item{mainLab}{A character representing the main label}

\item{legendLab}{A character representing the legend title.}

\item{cuDat}{Dataframe generated by \code{buildCUDat}.}

\item{plotVar}{A character value corresponding to PM in cuDat$vars (can be
either catch- or conservation-based).}

\item{group}{A character value that can take the values: \code{"mp", "om"}
and specifies along which categorical variable dot plots should be grouped.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates a radar plot. Note that PMs on different scales
(e.g. proportional abundance vs. absolute abundance) should not be plotted
together. If it shouldn't share a y-axis label, it shouldn't be in a radar
plot! Also cannot facet so should only be passed one OM and multiple MPs or
vice versa. Includes helper function coordRadar.
}
\examples{
trimDat <- agPlottingDF \%>\%
  dplyr::filter(var \%in\% c("ppnCULower", "ppnCUStable", "ppnFisheriesOpen")) \%>\%
  dplyr::mutate(var = factor(var))
plotRadar(trimDat, xLab = c("CUs\nLower BM", "CUs\nStable", "Fisheries\n Open"),
          plotVars = NULL, groupingVar = NULL, cu = FALSE,
          legendLab = "Proportion\nTAC in\nMixed Catch", axisSize = 13)

}
