% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppnErr.R
\name{ppnCatchErr}
\alias{ppnCatchErr}
\title{Generate log-normal error associated with catch data}
\usage{
ppnCatchErr(ppnCatchVec, tau)
}
\arguments{
\item{ppnCatchVec}{A numeric vector of proportions representing the true
relative contribution of each stock to total catch. \code{nStocks =
length(ppnAgeVec)}.}

\item{tau}{A numeric specifying the parameter that represents average var-
iation in proportions.}
}
\value{
Returns a numeric vector representing proportions for each stock.
}
\description{
Like the \code{ppnAgeErr} function, this function generates assigns multi-
variate logistic error (Schnute and Richards 1995, eqns.S.9 and S.10) to
proportions data. Here it is intended to represent error in assigning mixed
stock catches to specific stocks.
}
\examples{
ppnCatchVec <- c(0.2, 0.4, 0.3, 0.1)
tau <- 0.7
ppnCatchErr(ppnCatchVec, tau)

}
