% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedQuantiles.R
\name{qLow}
\alias{qLow}
\title{Calculate quantiles}
\usage{
qLow(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Functions to calculate lower and upper quantiles.
}
\details{
This is a generic wrapper function for quantile that defaults to specific values and removes
NAs.
}
\examples{
qLow(c(3, 12, NA, 1, 7, 25))
qHigh(c(3, 12, NA, 1, 7, 25))

}
