% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickLM.R
\name{quickLm}
\alias{quickLm}
\title{Linear model}
\usage{
quickLm(xVec, yVec)
}
\arguments{
\item{xVec, yVec}{Numeric vectors of equal length representing predictor and
response variables in linear model.}
}
\value{
Returns a list of model coefficients.
}
\description{
This function uses C++ to quickly fit linear models. Depends on fastLm.cpp.
}
\examples{
x <- rnorm(10, 0, 1)
y <- 2*x + 1
quickLm(x, y)

}
