% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_fraserRecMatrix.R
\docType{data}
\name{recMatrix}
\alias{recMatrix}
\title{Fraser River recruitment matrix}
\format{
A matrix with 61 rows representing year-specific observations and
11 columns representing CUs.
}
\source{
Fraser River Salmon Spawning Initiative (Ann-Marie Huang (DFO) personal
communication; January 2018)
}
\usage{
recMatrix
}
\description{
A matrix of recruitment estimates for 11 Fraser River sockeye salmon CUs
with time series extending to 1948. Recruitment is summed across all age
classes within a given brood year.
}
\keyword{datasets}
