% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverySimulator.R
\name{recoverySim}
\alias{recoverySim}
\title{Recovery simulator}
\usage{
recoverySim(
  simPar,
  cuPar,
  catchDat = NULL,
  srDat = NULL,
  variableCU = FALSE,
  makeSubDirs = TRUE,
  ricPars,
  larkPars = NULL,
  tamFRP = NULL,
  cuCustomCorrMat = NULL,
  erCorrMat = NULL,
  dirName,
  nTrials = 100,
  uniqueProd = TRUE,
  random = FALSE
)
}
\arguments{
\item{y}{TO BE DEFINED}
}
\value{
TO BE DEFINED.
}
\description{
Closed-loop simulation tool to assess management procedures and inform
Pacific salmon rebuilding strategies. Originally based on C. Holt's south
coast chum model (CSAS 2018). Simulation runs primed with observed SR data
(recDat input). The model includes data generation, variation in age
structure, survey design, and variable exploitation rules. OM uses Ricker
formulation \code{R=S(exp(a-bS))}. Additional details in changesToChumModel.md
in reports directory.
}
