% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_fraserRecAgePpns.R
\docType{data}
\name{recPpns}
\alias{recPpns}
\title{Proportion of recruits at age for Fraser River sockeye}
\format{
A DF with 1093 rows representing year-specific observations for each
CU.
}
\source{
Fraser River Salmon Spawning Initiative (Ann-Marie Huang (DFO) personal
communication; January 2018)
}
\usage{
recPpns
}
\description{
A dataframe of age-at-maturity proportions for 18 Fraser River sockeye salmon
CUs.
}
\keyword{datasets}
