% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockRecruitModels.R
\name{rickerModel}
\alias{rickerModel}
\title{Generate recruit abundance with Ricker model}
\usage{
rickerModel(S, a, b, error, rho = NULL, prevErr = NULL)
}
\arguments{
\item{S}{A numeric vector of spawner abundances.}

\item{a}{A numeric vector of alpha values, i.e. productivity at low spawner
abundance.}

\item{b}{A numeric vector of beta values, i.e. density dependence para-
meter.}

\item{error}{A numeric vector of recruitment deviations, typically generated
using \code{rmvnorm()} and relevant process variance estimates (sigma).}

\item{rho}{A numeric vector of rho values, i.e. AR1 coefficient.
outside of model using multivariate normal (or equivalent) distribution.}

\item{prevErr}{A numeric vector representing recruitment deviations from
previous brood year.}
}
\value{
Returns a list of R, a numeric representing recruit abundance, and
\code{errNext} which is used to generate subsequent process error (i.e. next
year's prevErr.
}
\description{
This function calculates recruitment from Ricker curve with AR(1) process
(according to Peterman et al. 2003; modified to take more recent parameter-
ization). Uses parameters from arima.mle (a, -b, sig, rho in log space) with
multivariate normally distributed errors. Note that internal \code{if}
statements prevent it from being vectorized so must be passed single values,
i.e. all vectors for inputs and outputs are length 1. Note that by default
prevErr and rho are NULL, resulting in a standard Ricker model.
}
\examples{
#Spawner and recruit values represent millions of fish, stock-recruit
parameters approximate those of Fraser River sockeye salmon Chilko CU.

#without autoregressive error
rickerModel(S = 1.1, a = 1.8, b = 1.2, error = 0.3)

#with autoregressive error
rickerModel(S = 1.1, a = 1.8, b = 1.2, error = 0.3, rho = 0.2,
prevErr = 0.7)

}
