% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_fraserSR.R
\docType{data}
\name{rickerParameters}
\alias{rickerParameters}
\title{Posterior estimates of Ricker stock recruit parameters}
\format{
A data frame with 19000 rows and 5 variables:
\describe{
  \item{stk}{stock id, approximately equivalent to a conservation unit}
  \item{alpha}{productivity parameter (i.e. recruits per spawner at low
  spawner abundances)}
  \item{beta0}{density dependence parameter}
  \item{sigma}{standard deviation of normally distributed process error with
  mean = 0}
  \item{deviance}{estimate of model fit equivalent to residual sum of
  squares for linear models}
}
}
\source{
Fraser River Salmon Spawning Initiative (Ann-Marie Huang (DFO) personal
communication; January 2018)
}
\usage{
rickerParameters
}
\description{
A dataset containing posterior estimates of stock recruit parameters for CU-
specific Ricker models fit to Fraser River sockeye salmon conservation units
(CUs).
}
\keyword{datasets}
