% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockRecruitModels.R
\name{rickerSurvModel}
\alias{rickerSurvModel}
\title{Generate recruit abundance with the Ricker Model with a marine survival co-variate}
\usage{
rickerSurvModel(S, a, b, error, ppnAges, gamma, mSurvAtAge)
}
\arguments{
\item{S}{A numeric of spawner abundance.}

\item{a}{A numeric of alpha value, i.e. productivity at low spawner
abundance.}

\item{b}{A numeric vector of beta value, i.e. density dependence para-
meter.}

\item{error}{A numeric recruitment deviation, typically generated
using \code{rmvnorm()} and relevant process variance estimates (sigma).}

\item{ppnAges}{A numeric vector of proportion of spawner abundance at age}

\item{gamma}{A numeric value represting the marine survival coefficient}

\item{mSurvAtAge}{A numeric vector representing marine survival covariates at age}
}
\value{
A list of recruitment at each age for the modelled brood year
(e.g., R2 = recruitment to age 2 from broodyear, R3 = recruitment to age 3, etc),
as well as recruitment for all return ages combined from the modelled brood year (RecBY)
}
\description{
This function calculates recruitment from Ricker curve with a marine survival covariate
that is specific to brood year added in.  Note that internal \code{if}
statements prevent it from being vectorized so must be passed single values,
i.e. all vectors for inputs and outputs are length 1.
}
\examples{
rickerSurvModel(S = 1000, a = 2.1, b = 0.00001, error = 0.8, ppnAges = c(0,0.83,0.17,0,0), gamma=0.4, mSurvAtAge=c(0,0.013, 0.015,0,0))

}
