% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRegime.R
\name{runRegime}
\alias{runRegime}
\title{compute length of regimes}
\usage{
runRegime(a1, a2, y, reglen = 10)
}
\arguments{
\item{a1}{Parameter value for the first regime. Value to be multiplied by the average parameter}

\item{a2}{Parameter value for the second regime.}

\item{y}{length of regime trend}

\item{reglen}{length of each regime, default is 10}
}
\value{
Returns a vector of length y with alternating sequence of regime scalars
}
\description{
This funncion computes the sequence of regime scalars based on used defines, 
low, high scallars as well as  regime length
}
\examples{
runRegime(.75,1.25,23,5)


}
