% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samSim.R
\docType{package}
\name{samSim}
\alias{samSim}
\title{samSim}
\description{
samSim is intended to evaluate management strategies for Pacific salmon
stocks while accounting for a range of uncertainties.
}
\author{
Cameron Freshwater <cameron.freshwater@dfo-mpo.gc.ca>
}
