% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBenchmarks.R
\name{sGenOptimum}
\alias{sGenOptimum}
\title{Optimize sGen}
\usage{
sGenOptimum(S, theta, sMSY)
}
\arguments{
\item{S}{A numeric of spawner abundances.}

\item{theta}{A numeric vector of Ricker stock recruit parameters: alpha,
beta, and sigma.}

\item{sMSY}{A numeric of spawner abundance estimated to result in maximum
sustainable yield.}
}
\value{
Returns a list of prt (estimated productivity), epsilon (the
difference between sMSY and productivity), the negative log likelihood
used to estimate sGen, and S.
}
\description{
This is a component function within sGenSolver that estimates the log-
likelihood of the relevant Ricker function.
}
