% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_synchList.R
\docType{data}
\name{synchList}
\alias{synchList}
\title{Simulated time series list}
\format{
A list with 9 elements. Elements 4:9 are arrays of time series data
with dimensions nYears x nCUs x nTrials.
\describe{
  \item{nameOM}{operating model name; medSynch represents an OM where CUs are
  moderately correlated with one another}
  \item{plotOrder}{a numeric that represents the order in which this scenario
  should be plotted relative to scenarios that were run simultaneously}
  \item{nPrime}{the number of years in the priming period; i.e. data from
  1:nPrime in the following arrays was observed, not simulated}
  \item{S}{spawner abundance}
  \item{recBY}{recruit abundance by brood year}
  \item{logRS}{log recruits per spawner}
  \item{recDev}{recruitment deviations; i.e. residuals from SR models fit to
  data or process error generated by the simulation}
  \item{migMort}{en route mortality}
  \item{singCatch}{catches in single stock fisheries}
}
}
\source{
Fraser River Salmon Spawning Initiative (Ann-Marie Huang (DFO) personal
communication; January 2018)
}
\usage{
synchList
}
\description{
A list containing example output data from a \code{recoverySimulator}
simulation run that is intended to be passed to \code{calcSynchMetrics}.
}
\keyword{datasets}
