% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleekTheme.R
\name{theme_sleekX}
\alias{theme_sleekX}
\title{Minimalist ggplot theme}
\usage{
theme_sleekX(
  base_size = 11,
  base_family = "",
  position = "standard",
  axisSize = 10,
  legendSize = 1,
  facetSize = 1.1
)
}
\arguments{
\item{base_size}{Defaults to 11 (don't change).}

\item{base_family}{Default font type (don't change).}

\item{position}{A character vector that can take the values \code{standard}
(default), \code{bottom}, \code{middle}, \code{top}, or \code{topWithX}.
Major differences between them are the presence of x-axis and facet labels.}

\item{axisSize}{A numeric representing font size for the axis labels;
defaults to 10.}

\item{legendSize}{A numeric scalar for the size of the legend; defaults to
1.}

\item{facetSize}{A numeric scalar for the relative size of facet labels.}
}
\value{
Returns a minimalist ggplot (i.e. no grid lines, more white and less
grey).
}
\description{
This function is a copy of S. Anderson's \code{ggsidekick} which is not
supported in the newest R version. Adds an argument for top, bottom, middle
for multipanel functionality.
}
