% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{trans}
\alias{trans}
\title{Utility transformation function}
\usage{
trans(alpha)
}
\arguments{
\item{tau}{Value for tau to test}

\item{logProps}{Log-transformed vector of mean proportion data}
}
\value{
Vector of simulated proportion data based on specified tau and means.
}
\description{
This is a component function within getTau used to calculate log proportions.
Originally written by B. Dorner and C. Holt so C. Freshwater unable to
fully verify documentation.
}
