% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroupRowwise.R
\name{ungroupRowwiseDF}
\alias{ungroupRowwiseDF}
\title{ungroupRowwiseDF}
\usage{
ungroupRowwiseDF(x)
}
\arguments{
\item{data}{Dataframe that is grouped rowwise.}
}
\value{
Returns a dataframe that is no longer grouped rowwise.
}
\description{
Simple helper function to undo rowwise grouping.
}
