% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_matprop.R
\docType{data}
\name{widedatprop}
\alias{widedatprop}
\title{Example of maturity proportions by brood year}
\format{
A dataframe with 5 columns used as example of maturity proportitions, extracted from RBT mature run
\describe{
  \item{brood_year}{brood year observations} 
  \item{2}{proportion of fish that mature at age 2 for that brood}
  \item{3}{proportion of fish that mature at age 3 for that brood}
  \item{4}{proportion of fish that mature at age 4 for that brood}
  \item{5}{proportion of fish that mature at age 5 for that brood}
 
}
}
\usage{
widedatprop
}
\description{
Example of maturity proportions by brood year
}
\keyword{datasets}
