% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericRecoverySimulator.r
\name{genericRecoverySim}
\alias{genericRecoverySim}
\title{Recovery Simulator - Generic Version}
\usage{
genericRecoverySim(
  simPar,
  cuPar,
  catchDat = NULL,
  srDat = NULL,
  variableCU = FALSE,
  makeSubDirs = TRUE,
  ricPars,
  larkPars = NULL,
  cuCustomCorrMat = NULL,
  erCorrMat = NULL,
  dirName,
  nTrials = 100,
  uniqueProd = TRUE,
  random = FALSE
)
}
\arguments{
\item{y}{TO BE DEFINED}
}
\value{
TO BE DEFINED
}
\description{
Closed-loop simulation tool to assess management procedures and inform
Pacific salmon rebuilding strategies.
}
\details{
Adapted from the original recoverySim() function developed to inform rebuilding
strategies for Fraser sockeye salmon and North Coast Chum (see Holt, C.A., Freshwater, C.,
Holt, K., and Huang, A.-M. 2020. A quantitative tool for evaluating rebuilding plans for
Pacific salmon. Can. Tech. Rep. Fish. Aquat. Sci. 3402: v + 26 p.;
https://waves-vagues.dfo-mpo.gc.ca/Library/40889385.pdf)

This version has been adapted to easily accomodate alternative life history types.
Additionally, the Total Allowable Mortality (TAM) HCR specific to Fraser sockeye fisheries,
and assocaited performance measures, has been removed.

Simulation runs primed with observed SR data
(recDat input). The model includes data generation, variation in age
structure, survey design, and variable exploitation rules.
}
