% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOutputList.R
\name{genOutputList}
\alias{genOutputList}
\title{Make a list of output data}
\usage{
genOutputList(dirName, subDirName = NULL, selectedCUs = NULL,
  agg = TRUE, aggTS = FALSE)
}
\arguments{
\item{dirName}{A character vector describing directory where output data saved.}

\item{subDirName}{A character vector describing subdirectory where output data saved.
Defaults to NULL.}

\item{selectedCUs}{A character vector of CUs that will be subsetted from aggregate. Defaults
to NULL.}

\item{agg}{Is the generated list summarizing aggregate or CU-specific data? Defaults to TRUE.}

\item{aggTS}{If agg == TRUE, is the generated list summarizing aggregate time series data?
Defaults to FALSE.}
}
\value{
Returns a list of subdirectory outputs within the directory.
}
\description{
This function is used to generate lists of summary data based on the outputs of recoverySim.
}
\details{
This is a function specific to the samSim package. Generally a directory represents either a
single operating model/management procedure with subdirectories representing representing OM/MPs
nested within creating multiple unique scenarios.
}
\examples{

}
