% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutputList.R
\name{genOutputList}
\alias{genOutputList}
\title{Generate list of outputs}
\usage{
genOutputList(
  dirName,
  subDirName = NULL,
  selectedCUs = NULL,
  agg = TRUE,
  aggTS = FALSE
)
}
\arguments{
\item{dirName}{A character vector representing the directory to pull data
from.}

\item{subDirName}{An optional character vector used to specify a single
subdirectory within \code{dirName} to pull data from.}

\item{selectedCUs}{An optional character vector of CU names used to subset
output data if CU-specific data are being pulled (i.e. \code{agg = FALSE}).}

\item{agg}{A logical stating whether aggregate (\code{TRUE}; default) or
CU-specific (\code{FALSE}) data should be pulled.}

\item{aggTS}{A logical stating whether aggregate summary performance metric
data (\code{FALSE}; default) or aggregate time series data (\code{TRUE})
should be pulled.}
}
\value{
Returns a list of output data of length equal to the number of unique
scenarios within a directory.
}
\description{
This function pulls data generated by \code{recoverySim} from a given
directory and binds it into a list. Each first order list element consists
of a single scenario (i.e. one OM and one MP). The second order list
elements depend on whether CU-specific, aggregate, or aggregate time series
data are extracted.
}
\examples{
TO BE COMPLETED

}
