% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{getTau}
\alias{getTau}
\title{Estimate tau}
\usage{
getTau(ppnMat, plotTaus = TRUE)
}
\arguments{
\item{ppnMat}{Matrix of proportion data to apply grid search for tau to.}

\item{plotTaus}{If true (default), tau-specific values from objective
function are plotted allowing users to visually identify minimum.}
}
\value{
A list containing a vector of tau-specific objective function values
(*objFun*) and a numeric representing the tau value that minimizes the
objective function (*bestTau*).
}
\description{
Main function used to estimate variation in multivariate proportion data.
Uses an objective function to minimize difference between sum of squares of
observed vs. simulated standard deviations for proportions data. Note that
objective function is constrained to sample between 0 and 3 at intervals of
0.1; however that should be sufficient for most proportion data.
}
\examples{

}
