% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{mvLogisticLL}
\alias{mvLogisticLL}
\title{MLE for multivariate logistic proportions}
\usage{
mvLogisticLL(agePropData)
}
\arguments{
\item{agePropData}{matrix of proportions at age}
}
\value{
List of parameter estimates for variance (tau) and mean proportions at age (pa).
}
\description{
This function provides MLE estimates for the multivariate logistic likelihood
function describe by Schunute and Richards 1995. Originally coded by Paul van Dam-Bates
}
