% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTradeoffs.R
\name{plotAgTradeoff}
\alias{plotAgTradeoff}
\title{Aggregate trade off plot}
\usage{
plotAgTradeoff(
  agDat,
  consVar = "medSpawners",
  catchVar = "medCatch",
  facet = "om",
  shape = NULL,
  hotColors = TRUE,
  showUncertainty = FALSE,
  legendLab = NULL,
  xLab = NULL,
  yLab = NULL,
  mainLab = NULL,
  axisSize = 14,
  dotSize = 4,
  lineSize = 1.25,
  legendSize = 14,
  scaleAxis = "free",
  facetLetter = FALSE
)
}
\arguments{
\item{agDat}{Dataframe generated by \code{buildAgDat}.}

\item{consVar}{A character value corresponding to a conservation based PM
in cuDat$vars.}

\item{catchVar}{A character value corresponding to a catch based PM in
cuDat$vars.}

\item{facet}{A character value that can take the values: \code{"mp", "om"}
and specifies along which categorical variable the plot should be faceted.}

\item{shape}{A character value that defaults to \code{NULL}, but can take
values \code{"mp"} or \code{"om"}, and specifies along which categorical
variable shapes should be plotted. Note maximum number of levels is 5.}

\item{hotColors}{A logical (default \code{TRUE}) that specifies whether
symbols should be filled with \code{viridis} palette or solid black.}

\item{showUncertainty}{A logical specifying whether whiskers for each
variables credible interval should be plotted.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{mainLab}{A character specifying a plot title (defaults to NULL).}

\item{scaleAxis}{A character vector that can take values `c("fixed", "free",
"free_x", "free_y")` that determines which axes, if any, have variable axes dimensions
across facets.}

\item{facetLetter}{A logical that determines facets are labelled with letters
for referencing in text.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates Kobe style plots for a series of conservation- and
catch-based performance metrics.
}
\examples{
plotAgTradeoff(agPlottingDF, consVar = "medSpawners", catchVar = "medCatch",
facet = "om", showUncertainty = TRUE,
legendLab = "Prop. TAC in mixed stock fishery", xLab = "Median Catch",
yLab = "Median Spawners")

}
