% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTradeoffs.R
\name{plotCUTradeoff}
\alias{plotCUTradeoff}
\title{CU-specific trade off plot}
\usage{
plotCUTradeoff(
  cuDat,
  consVar = "medSpawners",
  catchVar = "medCatch",
  facet = "cu",
  panel = NULL,
  showUncertainty = FALSE,
  hotColors = TRUE,
  legendLab = NULL,
  xLab = NULL,
  yLab = NULL,
  main = TRUE,
  axisSize = 14,
  dotSize = 4,
  lineSize = 1.25,
  legendSize = 14,
  freeY = TRUE,
  scaleAxis = "free"
)
}
\arguments{
\item{cuDat}{Dataframe generated by \code{buildCUDat}.}

\item{consVar}{A character value corresponding to a conservation based PM
in cuDat$vars.}

\item{catchVar}{A character value corresponding to a catch based PM in
cuDat$vars.}

\item{facet}{A character value that can take the values:
\code{"cu", "mp", "om"} and specifies along which categorical variable
the plot should be faceted.}

\item{panel}{A character value that can take the values:
\code{"mp", "om"} and specifies along which categorical variable new
pages in the output PDF will be generated.}

\item{showUncertainty}{A logical specifying whether whiskers for each
variable's credible interval should be plotted.}

\item{hotColors}{A logical (default \code{TRUE}) that specifies whether
symbols should be filled with \code{viridis} palette or grey scale.}

\item{legendLab}{A character representing the legend title.}

\item{xLab}{A character representing the x axis label.}

\item{yLab}{A character representing the y axis label.}

\item{main}{A logical specifying whether a plot title should be added.}

\item{scaleAxis}{A character vector that can take values `c("fixed", "free",
"free_x", "free_y")` that determines which axes, if any, have variable axes
dimensions across facets.}
}
\value{
Returns a ggplot object.
}
\description{
This function generates Kobe style plots for a series of conservation- and
catch-based performance metrics.
}
\examples{
plotCUTradeoff(cuPlottingDF, consVar = "medSpawners", catchVar = "medCatch",
facet = "cu", panel = "om", showUncertainty = FALSE,
legendLab = "Prop. TAC in mixed stock fishery", xLab = "Median Catch",
yLab = "Median Spawners", main = FALSE)

}
