% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuantile.R
\name{qHigh}
\alias{qHigh}
\title{Calculate upper quantile}
\usage{
qHigh(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns a numeric representing either the 5th or 95th percentile.
}
\description{
Two simple wrapper functions for \code{quantile()} that are used in plotting
functions. qLow defaults to the 5th percentile and qHigh to the 95th.
}
