% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuantile.R, R/fixedQuantiles.R
\name{qLow}
\alias{qLow}
\title{Calculate lower quantile}
\usage{
qLow(x)

qLow(x)
}
\arguments{
\item{x}{A numeric vector.}

\item{x}{A numeric vector.}
}
\value{
Returns a numeric representing either the 10th or 90th percentile.

A numeric vector.
}
\description{
Two simple wrapper functions for \code{quantile()} that are used in plotting
functions. qLow defaults to the 10th percentile and qHigh to the 90th.

Functions to calculate lower and upper quantiles.
}
\details{
This is a generic wrapper function for quantile that defaults to specific values and removes
NAs.
}
\examples{
qLow(c(3, 12, NA, 1, 7, 25))
qHigh(c(3, 12, NA, 1, 7, 25))

}
