% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_simPar.R
\docType{data}
\name{simParexample}
\alias{simParexample}
\title{Example of simPar input datafile}
\format{
A dataframe with 45 columns used as input to define simulation scenarios and MP:
\describe{
  \item{scenario}{scenario name. Used in subdirectories names as well as naming CUsrDat output file} 
  \item{nameOM}{operating model name. Used in naming simulation oyutput directories and many output files}
  \item{nameMP}{management procedure name, used in naming output files}
  \item{keyVar}{focal variable of the analysis; subjective since typically multiple variables will differ among scenarios, but should be a focal point of main figures. Currently can be one of the following arguments: `prodRegime`, `synch`, `expRate`, `ppnMix`, `sigma`, `endYear`, `adjustAge`, `mixOUSig`, `adjustForecast`, `adjustEnRoute`, `obsSig`, `obsMixCatch` (**NOTE these should eventually be defined explicitly**)}
  \item{plotOrder}{order in which grouped scenarios will be plotted (useful when keyVar is not an ordinal or numeric variable)}
  \item{species}{lower case species name (chum, sockeye and coho have been tested robustly; pink has not; chinook should be used with extreme caution since most stocks do not meet assumptions of the model- I am not sure this still holds)} 
  \item{simYears}{length of the simulation period (excluding priming period)}
  \item{harvContRule}{harvest control rule (`TAM`, `fixedER`, `genPA`)} 
  \item{benchmark}{biological benchmark used to assess conservation status (`stockRecruit`, `percentile`)}
  \item{canER}{total Canadian exploitation rate} 
  \item{usER}{ American exploitation rate (note can also be supplied as CU-specific value in `cuPars`)}
  \item{cvERSMU}{Annual variation in exploitation rate for the SMU}  
  \item{propMixHigh}{proportion of Canadian catch allocated to mixed-stock fisheries (can range from 0 to 1)}  
  \item{singleHCR}{single stock harvest control rule (`FALSE`, `retro`, `forecast`)}  
  \item{moveTAC}{if `TRUE` and single stock quota from low-abundance CUs is re-allocated to other CUs}   
  \item{prodRegime}{productivity regime (`low`, `lowStudT`, `med`, `studT`, `skew`, `skewT`, `decline`, `divergent`, `oneUp`,  `oneDown`, `high`)}   
  \item{prodPpnChange}{Scalar to modify productivy parameter, values >1 indicate increase linear trend and <1 declining linear trend}  
  \item{trendLength}{length of the trend in SR parameters in years, fo now trend happens in the beginning of the simulation time series}    
  \item{capRegime}{type of SR capacity (Smax) regime, current options are "linear", "regime"  any other option is treated as stable without warning}        
  \item{capPpnChange}{Scalar to modify Smax parameter, values >1 indicate increase linear trend and <1 declining linear trend}  
  \item{startYear}{indicates when a productivity decline (if specified by `prodRegime == "decline"`) should start -- not implemented}       
  \item{endYear}{indicates when a productivity decline (if specified by `prodRegime == "decline"`) should end}     
  \item{rho}{temporal autocorrelation coefficient in recruitment deviations}   
  \item{arSigTransform}{if `TRUE` estimates of sigma from input are transformed so that they account for temporal autocorrelation}   
  \item{biasCor}{if TRUE, apply bias correction to lognormal distribution (sig^2)/2}    
  \item{correlCU}{the correlation among CUs in recruitment deviations}   
  \item{corrMat}{if `TRUE` a custom correlation matrix is required to be passed as an input and is used to specify the covariance matrix for recruitment deviations}    
  \item{mu_logCovar1}{mean of lognormal distribution for annual SR covariate (e.g., marine survival)}    
  \item{sig_logCovar1}{log-normal variation for annual SR covariate}   
  \item{sampCU_coef1}{TRUE/FALSE indicating whether CU-specific coefficients for the SR covariate should be sampled}    
  \item{sigCU_coef1}{among-CU variation in coefficients for the SR covariate (only applied if sampCU_coef1 == TRUE)}      
  \item{tauCatch}{logistic variation in CU-specific catches}        
  \item{obsSig}{log-normal variation in spawner observation error}          
  \item{mixOUSig}{beta-distributed variation in mixed-stock fishery outcome uncertainty; input parameter represents the standard deviation used to calculate the location parameter}       
  \item{singOUSig}{beta-distributed variation in single-stock fishery outcome uncertainty; input parameter represents the standard deviation used to calculate the location parameter}      
  \item{obsMixCatchSig}{log-normal variation in mixed-stock catch observation error}  
  \item{obsSingCatchSig}{log-normal variation in single-stock catch observation error}  
  \item{obsAgeErr}{logistic variation in observed age error}       
  \item{lowCatchThresh}{lower aggregate catch target (used as a performance metric)}   
  \item{highCatchThresh}{upper aggregate catch target (used as a performance metric)}  
  \item{extinctThresh}{quasi-extinction threshold}    
  \item{adjustSig}{scalar on CU-specific sigmas (recruitment deviations)}       
  \item{adjustAge}{scalar on `tauCatch`}       
  \item{adjustForecast}{Scalar to adjust the sigma used to simulate population forecast}  
  \item{agePpnConst}{Logical. If TRUE set age proportion to be constant among CUs}     
}
}
\usage{
simParexample
}
\description{
Example of simPar input datafile
}
\keyword{datasets}
